Small code style changes disable convex hull disable leaflets cluster marker improve speed of first load with loading changes of minimap correct synchronization of events in player

+++ src/components/comparison/histogram/HistogramBox.js
@@ -45,9 +45,7 @@ const HistogramBox = ({dimensions, id}) => {
 
     const [openDelete, setOpen] = useState(false)
 
-    const setVisibility = (hide) => {
-        dispatch(changeVisibility(id, hide))
-    }
+    const setVisibility = (hide) => dispatch(changeVisibility(id, hide))
 
     const editEvent = () => {
         dispatch(setSelection(event))
@@ -55,15 +53,11 @@ const HistogramBox = ({dimensions, id}) => {
         navigate(`/selection`)
     }
 
-    const deleteId = () => {
-        setOpen(true)
-    }
+    const deleteId = () => setOpen(true)
 
     const handleClose = (deleteConfirmed) => {
         setOpen(false)
-        if (deleteConfirmed) {
-            dispatch(deleteEvent(id))
-        }
+        if (deleteConfirmed) dispatch(deleteEvent(id))
     }
 
     return (
